Fix two issues related to “Show Less”
Set a font by default when creating the attributed string version of the text. This prevents the code from later applying a font attribute to the entire string, potentially wiping out the custom font info for the “Show Less” part of the string.

Also, when creating the expanded version of the text, only add the show less link if the text will actually be truncated.